requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.5
cramjam==2.5.0
cycler==0.11.0
et-xmlfile==1.1.0
fastparquet==0.8.1
fonttools==4.36.0
fsspec==2022.7.1
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyarrow==9.0.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
tomli==2.0.1
